Generalize history serializer
